third_party/blink/web_tests/http/tests/cache/subresource-expiration-2-expected.txt

Cache-control: max-age=10, no-store; (result=Uncached expected=Uncached) PASS
Cache-control: max-age=10, no-cache="Set-Cookie"; (result=Cached expected=Cached) PASS
Cache-control: max-age=10, must-revalidate; (result=Cached expected=Cached) PASS
Age: 100; Cache-control: max-age=10; (result=Uncached expected=Uncached) PASS
Age: 1; Cache-control: max-age=10; (result=Cached expected=Cached) PASS
Cache-control: max-age=10; Pragma: no-cache; (result=Uncached expected=Uncached) PASS
Cache-control: max-age=0, private, must-revalidate, max-age=60; (result=Uncached expected=Uncached) PASS
Cache-control: max-age=60, private, must-revalidate, max-age=0; (result=Cached expected=Cached) PASS
Expires: [now-10s]; Cache-control: max-age=10; (result=Cached expected=Cached) PASS
Expires: [now+10s]; Cache-control: max-age=0; (result=Uncached expected=Uncached) PASS
Last-modified: [now-3600s]; [delay=1.5s] (result=Cached expected=Cached) PASS
Last-modified: [now-10s]; [delay=1.5s] (result=Uncached expected=Uncached) PASS
Last-modified: [now+10s]; (result=Uncached expected=Uncached) PASS
